logical code restructure : alerting servers
