+++ plugins/alerting/pkg/alerting/endpoints/v1/server.go
@@ -1,0 +1,246 @@
+package endpoints
+
+import (
+	"context"
+	"time"
+
+	"github.com/rancher/opni/pkg/alerting/shared"
+	"github.com/rancher/opni/pkg/alerting/storage"
+	"github.com/rancher/opni/pkg/alerting/storage/opts"
+	alertingv1 "github.com/rancher/opni/pkg/apis/alerting/v1"
+	corev1 "github.com/rancher/opni/pkg/apis/core/v1"
+	"github.com/rancher/opni/pkg/validation"
+	"github.com/samber/lo"
+	lop "github.com/samber/lo/parallel"
+
+	"golang.org/x/exp/slices"
+	"google.golang.org/protobuf/types/known/durationpb"
+	"google.golang.org/protobuf/types/known/emptypb"
+	"google.golang.org/protobuf/types/known/timestamppb"
+)
+
+var _ alertingv1.AlertEndpointsServer = (*EndpointServerComponent)(nil)
+
+func (e *EndpointServerComponent) CreateAlertEndpoint(ctx context.Context, req *alertingv1.AlertEndpoint) (*corev1.Reference, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+	newId := shared.NewAlertingRefId()
+	req.Id = newId
+	req.LastUpdated = timestamppb.Now()
+	if err := e.endpointStorage.Get().Put(ctx, newId, req); err != nil {
+		return nil, err
+	}
+	return &corev1.Reference{
+		Id: newId,
+	}, nil
+}
+
+func (e *EndpointServerComponent) GetAlertEndpoint(ctx context.Context, ref *corev1.Reference) (*alertingv1.AlertEndpoint, error) {
+	endp, err := e.endpointStorage.Get().Get(ctx, ref.Id)
+	if err != nil {
+		return nil, err
+	}
+	return endp, nil
+}
+
+func (e *EndpointServerComponent) UpdateAlertEndpoint(ctx context.Context, req *alertingv1.UpdateAlertEndpointRequest) (*alertingv1.InvolvedConditions, error) {
+	if err := unredactSecrets(ctx, e.endpointStorage.Get(), req.Id.Id, req.GetUpdateAlert()); err != nil {
+		return nil, err
+	}
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+
+	// TODO : this won't be possible in this server component
+	resp, err := e.notifications.ListRoutingRelationships(ctx, &emptypb.Empty{})
+	if err != nil {
+		return nil, err
+	}
+	refList := resp.GetInvolvedConditions(req.Id.Id)
+	if len(refList.Items) > 0 && !req.ForceUpdate {
+		return refList, nil
+	}
+	// force the new endpoint to preserve the original endpoint id
+	req.UpdateAlert.Id = req.Id.Id
+	req.UpdateAlert.LastUpdated = timestamppb.Now()
+	if err := e.endpointStorage.Get().Put(ctx, req.Id.Id, req.GetUpdateAlert()); err != nil {
+		return nil, err
+	}
+	return refList, nil
+}
+
+func (e *EndpointServerComponent) DeleteAlertEndpoint(ctx context.Context, req *alertingv1.DeleteAlertEndpointRequest) (*alertingv1.InvolvedConditions, error) {
+	existing, err := e.GetAlertEndpoint(ctx, req.Id)
+	if err != nil {
+		return nil, err
+	}
+	if existing == nil {
+		return &alertingv1.InvolvedConditions{}, nil
+	}
+
+	resp, err := e.notifications.ListRoutingRelationships(ctx, &emptypb.Empty{})
+	if err != nil {
+		return nil, err
+	}
+	refList := resp.GetInvolvedConditions(req.Id.Id)
+	if len(refList.Items) > 0 && !req.ForceDelete {
+		return refList, nil
+	}
+
+	if err := e.endpointStorage.Get().Delete(ctx, req.Id.Id); err != nil {
+		return nil, err
+	}
+
+	// TODO : this isn't available in this server component
+	lop.ForEach(refList.Items, func(condRef *corev1.Reference, _ int) {
+		// delete endpoint metadata from each condition
+		cond, err := e.conditionStorage.Get().Get(ctx, condRef.Id)
+		if err != nil {
+			return
+		}
+		if cond.AttachedEndpoints != nil && len(cond.AttachedEndpoints.Items) > 0 {
+			cond.AttachedEndpoints.Items = lo.Filter(cond.AttachedEndpoints.Items, func(item *alertingv1.AttachedEndpoint, _ int) bool {
+				return item.EndpointId != req.Id.Id
+			})
+		}
+		if err := e.conditionStorage.Get().Put(ctx, condRef.Id, cond); err != nil {
+			return
+		}
+	})
+
+	return refList, nil
+}
+func (e *EndpointServerComponent) ToggleNotifications(ctx context.Context, req *alertingv1.ToggleRequest) (*emptypb.Empty, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+
+	endp, err := e.endpointStorage.Get().Get(ctx, req.Id.Id, opts.WithUnredacted())
+	if err != nil {
+		return nil, err
+	}
+	if endp.GetProperties() == nil {
+		endp.Properties = map[string]string{}
+	}
+	if _, ok := endp.Properties[alertingv1.EndpointTagNotifications]; !ok {
+		endp.Properties[alertingv1.EndpointTagNotifications] = "true"
+	} else {
+		delete(endp.Properties, alertingv1.EndpointTagNotifications)
+	}
+	endp.LastUpdated = timestamppb.Now()
+	return &emptypb.Empty{}, e.endpointStorage.Get().Put(ctx, req.Id.Id, endp)
+}
+
+func (e *EndpointServerComponent) ListAlertEndpoints(
+	ctx context.Context,
+	req *alertingv1.ListAlertEndpointsRequest,
+) (*alertingv1.AlertEndpointList, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+	endpoints, err := e.endpointStorage.Get().List(ctx)
+	if err != nil {
+		return nil, err
+	}
+	items := []*alertingv1.AlertEndpointWithId{}
+	for _, endp := range endpoints {
+		items = append(items, &alertingv1.AlertEndpointWithId{
+			Id:       &corev1.Reference{Id: endp.Id},
+			Endpoint: endp,
+		})
+	}
+
+	slices.SortFunc(
+		items, func(i, j *alertingv1.AlertEndpointWithId) bool {
+			return i.Endpoint.Name < j.Endpoint.Name
+		},
+	)
+	return &alertingv1.AlertEndpointList{Items: items}, nil
+}
+
+func (e *EndpointServerComponent) TestAlertEndpoint(ctx context.Context, req *alertingv1.TestAlertEndpointRequest) (*alertingv1.TestAlertEndpointResponse, error) {
+	// lg := p.Logger.With("Handler", "TestAlertEndpoint")
+	if req.Endpoint == nil {
+		return nil, validation.Error("Endpoint must be set")
+	}
+	// if it has an Id it needs to be unredacted
+	if req.Endpoint.Id != "" {
+		unredactSecrets(ctx, e.endpointStorage.Get(), req.Endpoint.Id, req.Endpoint)
+	}
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+	details := &alertingv1.EndpointImplementation{
+		Title: "Test Alert Endpoint",
+		Body:  "Opni Alerting is sending you a test alert to verify your alert endpoint configuration.",
+	}
+
+	ephemeralId := shared.NewAlertingRefId()
+	createImpl := &alertingv1.FullAttachedEndpoints{
+		InitialDelay: durationpb.New(time.Duration(time.Second * 0)),
+		Items: []*alertingv1.FullAttachedEndpoint{
+			{
+				EndpointId:    ephemeralId,
+				AlertEndpoint: req.GetEndpoint(),
+				Details:       details,
+			},
+		},
+		ThrottlingDuration: durationpb.New(time.Duration(time.Second * 1)),
+		Details:            details,
+	}
+
+	// - create ephemeral dispatcher
+	router, err := e.routerStorage.Get().Get(ctx, shared.SingleConfigId)
+	if err != nil {
+		return nil, err
+	}
+
+	ns := "test"
+	if err := router.SetNamespaceSpec("test", ephemeralId, createImpl); err != nil {
+		return nil, err
+	}
+	// TODO : this won't be possible from this server component
+	go func() { // create, trigger, delete
+		ctx := e.ctx
+		e.opsNode.Get().SendManualSyncRequest(ctx, []string{shared.SingleConfigId}, e.routerStorage.Get())
+
+		_, err := e.notifications.TriggerAlerts(ctx, &alertingv1.TriggerAlertsRequest{
+			ConditionId: &corev1.Reference{Id: ephemeralId},
+			Namespace:   ns,
+			Annotations: map[string]string{
+				shared.OpniHeaderAnnotations: "Test notification",
+				shared.OpniBodyAnnotations:   "Admin has sent a test notification",
+			},
+			Labels: map[string]string{
+				ns: ephemeralId,
+			},
+		})
+		if err != nil {
+			e.logger.Errorf("Failed to trigger alert %s", err)
+		}
+
+		// - delete ephemeral dispatcher
+		if err := router.SetNamespaceSpec("test", ephemeralId, &alertingv1.FullAttachedEndpoints{
+			Items: []*alertingv1.FullAttachedEndpoint{},
+		}); err != nil {
+			return
+		}
+	}()
+
+	return &alertingv1.TestAlertEndpointResponse{}, nil
+}
+
+func unredactSecrets(
+	ctx context.Context,
+	store storage.EndpointStorage,
+	endpointId string,
+	endp *alertingv1.AlertEndpoint,
+) error {
+	unredacted, err := store.Get(ctx, endpointId, opts.WithUnredacted())
+	if err != nil {
+		return err
+	}
+	endp.UnredactSecrets(unredacted)
+	return nil
+}
